feat(lab): Nemotron ± ClawQL OpenRouter pair (no Anthropic)
Add openrouter/ baseline adapter and default GHA arms to
nemotron,nemotron-clawql with gpt-5.4-mini judge — OpenRouter-only.
